refactor: move BcrypterAdpter files to a folder
