+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:42121/jest?","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:33045/jest?","mongoDBName":"jest"}
