+++ src/infra/cryptography/bcrypt-adapter/bcrypt-adapter.ts
@@ -1,6 +1,6 @@
 import bcrypt from 'bcryptjs'
-import { HashComparer } from '../../data/protocols/cryptography/hash-comparer'
-import { Hasher } from '../../data/protocols/cryptography/hasher'
+import { HashComparer } from '../../../data/protocols/cryptography/hash-comparer'
+import { Hasher } from '../../../data/protocols/cryptography/hasher'
 
 export class BcryptAdapter implements Hasher, HashComparer {
   private readonly salt: number
